# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get current authenticated user with role and progress summary
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
    patch:
      operationId: updateMe
      tags: [me]
      summary: Update profile (name, avatar, bio)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMeBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /me/billing:
    get:
      operationId: getBilling
      tags: [me]
      responses:
        "200":
          description: Billing settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Billing" }
    patch:
      operationId: updateBilling
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBillingBody" }
      responses:
        "200":
          description: Updated billing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Billing" }
...
  /me/invoices/{invoiceId}/download:
    parameters:
      - { name: invoiceId, in: path, required: true, schema: { type: string } }
      - { name: source, in: query, required: false, schema: { type: string, enum: [stripe, local] } }
    get:
      operationId: downloadInvoice
      tags: [me]
      responses:
        "200":
          description: Invoice PDF stream
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "302": { description: Redirect to Stripe-hosted PDF }
        "404": { description: Not found }
...
components:
  schemas:
...
    Role:
      type: string
      enum: [super_admin, admin, team, student]
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
    MeResponse:
      type: object
      required: [user, totalLessons, completedLessons]
      properties:
        user: { $ref: "#/components/schemas/User" }
        totalLessons: { type: integer }
        completedLessons: { type: integer }
    UpdateMeBody:
      type: object
      properties:
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
    Billing:
      type: object
      required: [plan, status]
      properties:
        plan: { type: string }
        status: { type: string }
        billingEmail: { type: string, nullable: true }
        company: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
        nextInvoiceDate: { type: string, format: date-time, nullable: true }
        invoices:
          type: array
          items: { $ref: "#/components/schemas/Invoice" }
    Invoice:
      type: object
      required: [id, date, amount, status]
      properties:
        id: { type: string }
        date: { type: string, format: date-time }
        amount: { type: string }
        status: { type: string }
        description: { type: string }
...
    UpdateBillingBody:
      type: object
      properties:
        billingEmail: { type: string, nullable: true }
        company: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
...